+++ src/components/scenes/WalletDetailsScene.tsx
@@ -22,6 +22,7 @@ import { getExchangeDenomByCurrencyCode } from '../../selectors/DenominationSele
 import { getExchangeRate } from '../../selectors/WalletSelectors'
 import { FooterRender } from '../../state/SceneFooterState'
 import { useSceneScrollHandler } from '../../state/SceneScrollState'
+import { config } from '../../theme/appConfig'
 import { useDispatch, useSelector } from '../../types/reactRedux'
 import { NavigationBase, RouteProp, WalletsTabSceneProps } from '../../types/routerTypes'
 import { coinrankListData, infoServerData } from '../../util/network'
@@ -65,6 +66,8 @@ function WalletDetailsComponent(props: Props) {
   const tokenId = checkToken(route.params.tokenId, wallet.currencyConfig.allTokens)
   const { pluginId } = wallet.currencyInfo
   const { currencyCode, displayName } = tokenId == null ? wallet.currencyInfo : wallet.currencyConfig.allTokens[tokenId]
+
+  const educationCards = (infoServerData.rollup?.assetInfoCards ?? {})[`${pluginId}${tokenId == null ? '' : `_${tokenId}`}`] ?? []
 
   // State:
   const scrollViewRef = React.useRef<AnimatedScrollView>(null)
@@ -319,6 +322,13 @@ function WalletDetailsComponent(props: Props) {
               <BuyCrypto navigation={navigation as NavigationBase} wallet={wallet} tokenId={tokenId} />
             )}
           </View>
+          {educationCards.length === 0 ? null : (
+            <>
+              <DividerLineUi4 extendRight />
+              <SectionHeaderUi4 leftTitle={config.appName === 'Edge' ? lstrings.edge_ucation : lstrings.education} />
+              <InfoCardCarousel enterAnim={fadeInDown10} cards={educationCards} navigation={navigation as NavigationBase} screenWidth={screenWidth} />
+            </>
+          )}
         </Reanimated.ScrollView>
       )}
     </SceneWrapper>